+++ SkillModifiers/teleport.lua
@@ -127,7 +127,7 @@ Initialize(function()
     gm.surface_free(my_surface)
 end)
 
-local teleport = SkillModifierManager.register_modifier("teleport", 250)
+local teleport = SkillModifierManager.register_modifier("teleport", 124)
 teleport:set_add_func(function(data, modifier_index, teleport_id)
     data:add_post_drop_callback(function(actor, skill_params)
         gm.actor_skill_set(actor, skill_params.slot_index, skill.value)